+++ debug/example/cachestorage/cachestorage.js
@@ -10,23 +10,19 @@ const onReady = () => {
   const dataString = "홍길동";
   const dataArray = ["유관순", "이순신", "홍길동"];
   // 캐시 데이터 체크 및 저장 및 데이터 가져오기
-  const cachedData = (_a = window.cacheStorage._get('mc_object')) !== null && _a !== void 0 ? _a : (() => {
-    window.cacheStorage._set('mc_object', data, 60);
-    return window.cacheStorage._get('mc_object');
-  })();
+  const cachedData = (_a = window.cacheStorage._get('mc_object')) !== null && _a !== void 0 ? _a : window.cacheStorage._set('mc_object', data, 60)._get('mc_object');
   Log.d('cache', 'mc_object', cachedData);
   // string
-  const cachedString = (_b = window.cacheStorage._get('mc_string')) !== null && _b !== void 0 ? _b : (() => {
-    window.cacheStorage._set('mc_string', dataString, 10);
-    return window.cacheStorage._get('mc_string');
-  })();
+  const cachedString = (_b = window.cacheStorage._get('mc_string')) !== null && _b !== void 0 ? _b : window.cacheStorage._set('mc_string', dataString, 10)._get('mc_string');
   Log.d('cache string', 'mc_string', cachedString);
   // array
-  const cachedArray = (_c = window.cacheStorage._get('mc_array')) !== null && _c !== void 0 ? _c : (() => {
-    window.cacheStorage._set('mc_array', dataArray, 10);
-    return window.cacheStorage._get('mc_array');
-  })();
+  const cachedArray = (_c = window.cacheStorage._get('mc_array')) !== null && _c !== void 0 ? _c : window.cacheStorage._set('mc_array', dataArray, 10)._get('mc_array');
   Log.d('cache array', 'mc_array', cachedArray);
+  // 비동기 
+  window.cacheStorage._getAsync('mc_array').
+  then((data) => {
+    Log.d('_getAsync', data);
+  });
   // 캐시 삭제
   // window.cacheStorage._delete('mc');
   // 캐시 전체 비우기