+++ src/example/cachestorage/cachestorage.ts
@@ -16,26 +16,23 @@ const onReady = () : void =>
     const dataArray = ["유관순","이순신","홍길동"];
 
     // 캐시 데이터 체크 및 저장 및 데이터 가져오기
-    const cachedData = window.cacheStorage._get('mc_object') ?? (() => {
-        window.cacheStorage._set('mc_object', data, 60);
-        return window.cacheStorage._get('mc_object');
-    })();
+    const cachedData = window.cacheStorage._get('mc_object') ?? window.cacheStorage._set('mc_object', data, 60)._get('mc_object');
     Log.d('cache','mc_object', cachedData);
 
 
     // string
-    const cachedString = window.cacheStorage._get('mc_string') ?? (() => {
-        window.cacheStorage._set('mc_string', dataString, 10);
-        return window.cacheStorage._get('mc_string');
-    })();
+    const cachedString = window.cacheStorage._get('mc_string') ?? window.cacheStorage._set('mc_string', dataString, 10)._get('mc_string');
     Log.d('cache string','mc_string', cachedString);
 
     // array
-    const cachedArray = window.cacheStorage._get('mc_array') ?? (() => {
-        window.cacheStorage._set('mc_array', dataArray, 10);
-        return window.cacheStorage._get('mc_array');
-    })();
+    const cachedArray = window.cacheStorage._get('mc_array') ?? window.cacheStorage._set('mc_array', dataArray, 10)._get('mc_array');
     Log.d('cache array','mc_array', cachedArray);
+
+    // 비동기 
+    window.cacheStorage._getAsync('mc_array')
+    .then(data =>{
+        Log.d('_getAsync', data);
+    });
 
     // 캐시 삭제
     // window.cacheStorage._delete('mc');